+++ R/smooth-spline.r
@@ -2,8 +2,8 @@
 #'
 #' Spline interpolation uses [stats::spline()] to interpolate between existing
 #' vertices using piecewise cubic polynomials. The `x` and `y` coordinates are
-#' interpolated independently. The feature curve will always pass through the
-#' vertices of the original feature.
+#' interpolated independently. The curve will always pass through the vertices
+#' of the original feature.
 #'
 #' This function works on matrices of points and is generally not called
 #' directly. Instead, use [smooth()] with `method = "spline"` to apply this
@@ -11,7 +11,7 @@
 #'
 #' @param x numeric matrix; 2-column matrix of coordinates.
 #' @param wrap logical; whether the coordinates should be wrapped at the ends,
-#'   as for polygons and closed lines.
+#'   as for polygons and closed lines, to ensure a smooth edge.
 #' @param vertex_factor double; the proportional increase in the number of
 #'   vertices in the smooth curve. For example, if the original curve has 100
 #'   points, a value of `2.5` will yield a new smoothed curve with 250 points.
@@ -20,10 +20,10 @@
 #'
 #' @return A matrix with the coordinates of the smoothed curve.
 #' @references The spline method was inspired by the following StackExchange
-#'   posts:
+#'   answers:
 #'
-#'   - [Create polygon from set of points distributed](https://stackoverflow.com/questions/26087772/26089377)
-#'   - [Smoothing polygons in contour map?](https://gis.stackexchange.com/questions/24827/24929)
+#'   - [Create polygon from set of points distributed](https://stackoverflow.com/a/26089377/3591386)
+#'   - [Smoothing polygons in contour map?](https://gis.stackexchange.com/a/24929/26661)
 #' @seealso [smooth()]
 #' @export
 #' @examples