+++ R/smooth.r
@@ -6,9 +6,9 @@
 #' @param x spatial features; lines or polygons from either the sf or sp
 #'   packages.
 #' @param method character; specifies the type of smoothing method to use.
-#'   Possible methods are: `"chaikin"`, `"spline"`, and `"densify"`. Each method
-#'   has one or more parameters specifying the amount of smoothing to perform.
-#'   See Details for descriptions.
+#'   Possible methods are: `"chaikin"`, `"ksmooth"`, `"spline"`, and
+#'   `"densify"`. Each method has one or more parameters specifying the amount
+#'   of smoothing to perform. See Details for descriptions.
 #' @param ... additional arguments specifying the amount of smoothing, passed on
 #'   to the specific smoothing function, see Details below.
 #'
@@ -22,6 +22,18 @@
 #'   the next point and one 1/4 of the way to the previous point. Smoothing
 #'   parameters:
 #'     - `refinements`: number of corner cutting iterations to apply.
+#'   - [smooth_ksmooth()]: kernel smoothing via the [stats::ksmooth()] function.
+#'   This method first calls [smooth_densify()] to densify the feature, then
+#'   applies Gaussian kernel regression to smooth the resulting points.
+#'   Smoothing parameters:
+#'     - `n`: number of times to split each line segment in the densification
+#'     step. Ignored if `max_distance` is specified.
+#'     - `max_distance`: the maximum distance between vertices in the resulting
+#'     features for the densification step. This is the Euclidean distance and
+#'     not the great circle distance.
+#'     - `bandwidth`: the bandwidth of the Gaussian kernel. Larger bandwidths
+#'     will result in more smoothing. If no value is supplied, the bandwidth is
+#'     estimated from the data.
 #'   - [smooth_spline()]: spline interpolation via the [stats::spline()]
 #'   function. This method interpolates between existing vertices and can be
 #'   used when the resulting smoothed feature should pass through the vertices
@@ -33,7 +45,7 @@
 #'     - `n`: number of vertices in each smoothed feature.
 #'   - [smooth_densify()]: densification of vertices for lines and polygons.
 #'   This is not a true smoothing algorithm, rather new vertices are added to
-#'   each line seqment via linear interpolation. Densification parameters:
+#'   each line segment via linear interpolation. Densification parameters:
 #'     - `n`: number of times to split each line segment. Ignored if
 #'     `max_distance` is specified.
 #'     - `max_distance`: the maximum distance between vertices in the resulting
@@ -42,52 +54,63 @@
 #'
 #' @return A smoothed polygon or line in the same format as the input data.
 #' @references See specific smoothing function help pages for references.
-#' @seealso [smooth_chaikin()] [smooth_spline()] [smooth_densify()]
+#' @seealso [smooth_chaikin()] [smooth_ksmooth()] [smooth_spline()]
+#'   [smooth_densify()]
 #' @export
 #' @examples
 #' library(sf)
 #' # compare different smoothing methods
 #' # polygons
 #' par(mar = c(0, 0, 0, 0), oma = c(4, 0, 0, 0), mfrow = c(3, 3))
-#' p_smooth_spline <- smooth(jagged_polygons, method = "spline")
 #' p_smooth_chaikin <- smooth(jagged_polygons, method = "chaikin")
+#' p_smooth_ksmooth <- smooth(jagged_polygons, method = "ksmooth")
+#' p_smooth_spline <- smooth(jagged_polygons, method = "spline")
 #' for (i in 1:nrow(jagged_polygons)) {
 #'   plot(st_geometry(p_smooth_spline[i, ]), col = NA, border = NA)
 #'   plot(st_geometry(jagged_polygons[i, ]), col = "grey40", border = NA, add = TRUE)
 #'   plot(st_geometry(p_smooth_chaikin[i, ]), col = NA, border = "#E41A1C", lwd = 2, add = TRUE)
+#'   plot(st_geometry(p_smooth_ksmooth[i, ]), col = NA, border = "#4DAF4A", lwd = 2, add = TRUE)
 #'   plot(st_geometry(p_smooth_spline[i, ]), col = NA, border = "#377EB8", lwd = 2, add = TRUE)
 #' }
 #' par(fig = c(0, 1, 0, 1), oma = c(0, 0, 0, 0), new = TRUE)
 #' plot(0, 0, type = "n", bty = "n", xaxt = "n", yaxt = "n", axes = FALSE)
-#' legend("bottom", legend = c("chaikin", "spline"), col = c("#E41A1C", "#377EB8"),
+#' legend("bottom", legend = c("chaikin", "ksmooth", "spline"),
+#'        col = c("#E41A1C", "#4DAF4A", "#377EB8"),
 #'        lwd = 2, cex = 2, box.lwd = 0, inset = 0, horiz = TRUE)
 #'
 #' # lines
 #' par(mar = c(0, 0, 0, 0), oma = c(4, 0, 0, 0), mfrow = c(3, 3))
-#' l_smooth_spline <- smooth(jagged_lines, method = "spline")
 #' l_smooth_chaikin <- smooth(jagged_lines, method = "chaikin")
+#' l_smooth_ksmooth <- smooth(jagged_lines, method = "ksmooth")
+#' l_smooth_spline <- smooth(jagged_lines, method = "spline")
 #' for (i in 1:nrow(jagged_lines)) {
 #'   plot(st_geometry(l_smooth_spline[i, ]), col = NA)
 #'   plot(st_geometry(jagged_lines[i, ]), col = "grey20", lwd = 3, add = TRUE)
-#'   plot(st_geometry(l_smooth_spline[i, ]), col = "#E41A1C", lwd = 2, lty = 2, add = TRUE)
-#'   plot(st_geometry(l_smooth_chaikin[i, ]), col = "#377EB8", lwd = 2, lty = 2, add = TRUE)
+#'   plot(st_geometry(l_smooth_chaikin[i, ]), col = "#E41A1C", lwd = 2, lty = 2, add = TRUE)
+#'   plot(st_geometry(l_smooth_ksmooth[i, ]), col = "#4DAF4A", lwd = 2, lty = 2, add = TRUE)
+#'   plot(st_geometry(l_smooth_spline[i, ]), col = "#377EB8", lwd = 2, lty = 2, add = TRUE)
 #' }
 #' par(fig = c(0, 1, 0, 1), oma = c(0, 0, 0, 0), new = TRUE)
 #' plot(0, 0, type = "n", bty = "n", xaxt = "n", yaxt = "n", axes = FALSE)
-#' legend("bottom", legend = c("chaikin", "spline"), col = c("#E41A1C", "#377EB8"),
+#' legend("bottom", legend = c("chaikin", "smooth", "spline"),
+#'        col = c("#E41A1C", "#4DAF4A", "#377EB8"),
 #'        lwd = 2, cex = 2, box.lwd = 0, inset = 0, horiz = TRUE)
-smooth <- function(x, method = c("chaikin", "spline", "densify"), ...) {
+smooth <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
+                   ...) {
   UseMethod("smooth")
 }
 
 #' @export
-smooth.sfg <- function(x, method = c("chaikin", "spline", "densify"), ...) {
+smooth.sfg <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
+                       ...) {
   method <- match.arg(method)
   # choose smoother
-  if (method == "spline") {
-    smoother <- smooth_spline
-  } else if (method == "chaikin") {
+  if (method == "chaikin") {
     smoother <- smooth_chaikin
+  } else if (method == "ksmooth") {
+    smoother <- smooth_ksmooth
+  } else if (method == "spline") {
+    smoother <- smooth_spline
   } else if (method == "densify") {
     smoother <- smooth_densify
   } else {
@@ -120,7 +143,8 @@ smooth.sfg <- function(x, method = c("chaikin", "spline", "densify"), ...) {
 }
 
 #' @export
-smooth.sfc <- function(x, method = c("chaikin", "spline", "densify"), ...) {
+smooth.sfc <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
+                       ...) {
   method <- match.arg(method)
   for (i in seq_along(x)) {
     x[[i]] <- smooth(x[[i]], method = method, ...)
@@ -129,14 +153,17 @@ smooth.sfc <- function(x, method = c("chaikin", "spline", "densify"), ...) {
 }
 
 #' @export
-smooth.sf <- function(x, method = c("chaikin", "spline", "densify"), ...) {
+smooth.sf <- function(x, method = c("chaikin", "ksmooth", "spline", "densify"),
+                      ...) {
   method <- match.arg(method)
   sf::st_geometry(x) <- smooth(sf::st_geometry(x), method = method, ...)
   x
 }
 
 #' @export
-smooth.Spatial <- function(x, method = c("chaikin", "spline", "densify"), ...) {
+smooth.Spatial <- function(x, method = c("chaikin", "ksmooth", "spline",
+                                         "densify"),
+                           ...) {
   if (!requireNamespace("sp", quietly = TRUE)) {
     stop("Install the sp package to smooth sp features.")
   }