+++ tests/testthat/test_smooth.r
@@ -3,23 +3,33 @@ library(sf)
 
 test_that("smooth() methods work", {
   # polygons
-  s <- smooth(jagged_polygons[1:6, ], method = "spline")
-  # change precision to fix some floating point issues on windows
-  s <- st_set_precision(s, 1e6)
-  expect_true(all(st_is_valid(s)))
+  # chaking
   s <- smooth(jagged_polygons, method = "chaikin")
   # change precision to fix some floating point issues on windows
   s <- st_set_precision(s, 1e6)
   expect_true(all(st_is_valid(s)))
+  # ksmooth
+  s <- smooth(jagged_polygons[1:6, ], method = "ksmooth")
+  # change precision to fix some floating point issues on windows
+  s <- st_set_precision(s, 1e6)
+  expect_true(all(st_is_valid(s)))
+  # spline
+  s <- smooth(jagged_polygons[1:6, ], method = "spline")
+  # change precision to fix some floating point issues on windows
+  s <- st_set_precision(s, 1e6)
+  expect_true(all(st_is_valid(s)))
+  # densify
   s <- smooth(jagged_polygons, method = "densify")
   # change precision to fix some floating point issues on windows
   s <- st_set_precision(s, 1e6)
   expect_true(all(st_is_valid(s)))
 
   # lines
-  sl <- smooth(jagged_lines, method = "spline")
-  expect_true(all(st_is_valid(s)))
   sl <- smooth(jagged_lines, method = "chaikin")
+  expect_true(all(st_is_valid(s)))
+  sl <- smooth(jagged_lines, method = "ksmooth")
+  expect_true(all(st_is_valid(s)))
+  sl <- smooth(jagged_lines, method = "spline")
   expect_true(all(st_is_valid(s)))
   sl <- smooth(jagged_lines, method = "densify")
   expect_true(all(st_is_valid(s)))
@@ -30,6 +40,15 @@ test_that("smooth() methods work", {
   s_r2 <- smooth(p, method = "chaikin", refinements = 1)
   s_r5 <- smooth(p, method = "chaikin", refinements = 5)
   expect_lt(nrow(s_r2[[1]]), nrow(s_r5[[1]]))
+  # ksmooth
+  s_n2 <- smooth(p, method = "ksmooth", n = 2)
+  s_n3 <- smooth(p, method = "ksmooth", n = 3)
+  expect_lt(nrow(s_n2[[1]]), nrow(s_n3[[1]]))
+  expect_equal(nrow(s_n2[[1]]), 2 * (nrow(p[[1]]) - 1) + 1)
+  expect_equal(nrow(s_n3[[1]]), 3 * (nrow(p[[1]]) - 1) + 1)
+  s_md1 <- smooth(p, method = "ksmooth", max_distance = 0.1)
+  s_md05 <- smooth(p, method = "ksmooth", max_distance = 0.05)
+  expect_lt(nrow(s_md1[[1]]), nrow(s_md05[[1]]))
   # spline
   s_n100 <- smooth(p, method = "spline", n = 100)
   s_n1000 <- smooth(p, method = "spline", n = 1000)
@@ -71,19 +90,25 @@ test_that("smooth() works for different input formats", {
 test_that("smooth() preserves holes", {
   p <- jagged_polygons$geometry[[5]]
   expect_true(st_is_valid(smooth(p, method = "chaikin")))
+  expect_true(st_is_valid(smooth(p, method = "ksmooth")))
   expect_true(st_is_valid(smooth(p, method = "spline")))
+  expect_true(st_is_valid(smooth(p, method = "densify")))
   expect_equal(length(p), length(smooth(p)))
 })
 
 test_that("smooth() preserves multipart features", {
   p <- jagged_polygons$geometry[[7]]
   expect_true(st_is_valid(smooth(p, method = "chaikin")))
+  expect_true(st_is_valid(smooth(p, method = "ksmooth")))
   expect_true(st_is_valid(smooth(p, method = "spline")))
+  expect_true(st_is_valid(smooth(p, method = "densify")))
   expect_equal(length(p), length(smooth(p)))
 
   l <- jagged_lines$geometry[[8]]
   expect_true(st_is_valid(smooth(l, method = "chaikin")))
+  expect_true(st_is_valid(smooth(l, method = "ksmooth")))
   expect_true(st_is_valid(smooth(l, method = "spline")))
+  expect_true(st_is_valid(smooth(l, method = "densify")))
   expect_equal(length(l), length(smooth(l)))
 })