+++ NEWS.md
@@ -1,5 +1,9 @@
 # smoothr 0.0.1.9000
 
+- Kernel smoothing implemented in `smooth_ksmooth()`
+- Densification of spatial lines and polygons implemented with `densify()`
+- Raster example dataset added
+
 # smoothr 0.0.1
 
 - Initial release of smoothr